Move translation file to `src` folder
