+++ src/pages/index.tsx
@@ -1,16 +1,16 @@
 import React from "react";
 
-import "../../translation";
-import Hero from '../components/Hero';
+import "../translation";
+import Hero from "../components/Hero";
 
-import 'normalize.css';
-import '../styles/main.scss';
+import "normalize.css";
+import "../styles/main.scss";
 import LanguageSwitcher from "../components/LanguageSwitcher";
 
 const Index = () => (
   <div className="blue-theme">
-    <LanguageSwitcher/>
-    <Hero/>
+    <LanguageSwitcher />
+    <Hero />
   </div>
 );
